\name{dataset1}
\alias{dataset1}
\docType{data}
\title{
Simulated data set
}
\description{
This is a dataset used in the examples.
}
\usage{data(dataset1)}
\format{
  A data frame with 180 observations on the following 5 variables.
  \describe{
    \item{\code{X1}}{a numeric vector}
    \item{\code{X2}}{a numeric vector}
    \item{\code{X3}}{a numeric vector}
    \item{\code{X4}}{a numeric vector}
    \item{\code{X5}}{a numeric vector}
  }
}

\examples{
data("dataset1")
## maybe str(dataset1) ; plot(dataset1) ...
}
\keyword{datasets}
